Live plot: Move plot viewer in separate file
